+++ changelog.md
@@ -8,6 +8,7 @@ The BounCA change history
 * Made Subject Alternative Name mandatory for client, server certificate as browsers require them
 * Add revoke prefix to revoked certificates name
 * Don't display revoked certificates in overview screen
+* App token for downloading CRL certificate
 
 ## [0.3.0] - Release 2022-02-14
 * Added resend e-mail button if user has not verified email yet